+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - knowledgium material
 
+### Fixed
+
+- JEI integration for Automata recipes
+
 ## [0.0.2] - 2021-07-29
 
 ### Changed