+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- knowledgium material
-- recipe registry [#15]
+- Knowledgium material
+- [Recipe registry](https://github.com/SirEdvin/ProgressivePeripherals/issues/15)
 
 ### Fixed