+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2021-07-29
+
 ### Added
 
 - Knowledgium material
@@ -15,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - JEI integration for Automata recipes
+- Problem with `getLevel` abstract
 
 ## [0.0.2] - 2021-07-29